Move `::test_last_modified_time_is_updated()
